feature: add OnError hook (#398)
* Add OnError hook

Adds a new hook that is called when a rest request returns an error.
This is called when the err returned by client.execute is non-nil, and
it is only called after all retry attempts have been exhausted.

This feature is intended to support logging, metrics, tracing on request
errors that otherwise cannot be captured with OnAfterResponse, such as
TLS Connection errors. It also provides a way to report response errors
only once after retries have been attempted, to avoid emitting error
logs when a request ultimately succeeds.

* Wrap OnError errors with *ResponseError

If a request results in an error which also has a response from the
server, it will be wrapped with responseError so that the response is
available to the OnError hook. The hook can use a type assertion or
errors.As to get the response.

client.OnError(func(err error) {
  if v, ok := err.(interface{ Response() *resty.Response }); ok {
    resp := v.Response()
  }
  // Log the error, increment a metric, etc...
})

* Include *Request in ErrorHook func

* Add more test coverage

* Update readme for OnError hook

* Trigger OnError hook on SRV lookup failure